server/requirements.txt
Django==3.0.4
dj-database-url==0.5.0   # auto database configuration
Fabric3==1.14.post1   # fabfile.py instrument
gunicorn==20.0.4   # Python WSGI HTTP server
psycopg2-binary==2.8.4   # PostgreSQL adapter
Pillow==7.1.2   # for ImageField
djangorestframework==3.11.0   # for REST API
djangorestframework-simplejwt==4.4.0   # for JWT support
django-filter==2.3.0   # for DRF filters
django-redis==4.11.0   # Redis

### For codestyle and testing
flake8==3.7.9   # linter
pylint==2.4.4   # linter
isort==4.3.21   # imports order
pylint-django==2.0.15   # linter
factory-boy==2.12.0   # for creating test`s models
